Move getAmountToPay
