+++ packages/arb-token-bridge-ui/src/components/TransferPanel/TransferPanel.tsx
@@ -91,10 +91,11 @@ import { useError } from '../../hooks/useError'
 import { isLifiRoute, useRouteStore } from './hooks/useRouteStore'
 import { LifiTransferStarter } from '@/token-bridge-sdk/LifiTransferStarter'
 import { getFromAndToTokenAddresses } from './LifiSettings'
-import { getAmountLoss, getAmountToPay } from './HighSlippageWarningDialog'
+import { getAmountLoss } from './HighSlippageWarningDialog'
 import { useLifiMergedTransactionCacheStore } from '../../hooks/useLifiMergedTransactionCacheStore'
 import { getStepTransaction } from '@lifi/sdk'
 import { isValidTransactionRequest } from '../../util/isValidTransactionRequest'
+import { getAmountToPay } from './useTransferReadiness'
 
 const signerUndefinedError = 'Signer is undefined'
 const transferNotAllowedError = 'Transfer not allowed'
